requirements.txt
fastapi
line-bot-sdk
python-dotenv
mangum
openai
scikit-learn
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client